fixed morebugs in latch-free

+++ hashtable/LockRequestLinkedList.h
@@ -30,7 +30,7 @@ private:
   int size_to_req;
   TLinkedList<MemoryChunk<TNode<LockRequest> > >* memory_list;
   LockPool * lock_pool;
-  int allocating = 0;
+  uint64_t allocating = 0;
 };
 
 #endif
